Extracted tangent calculation from Mesh into util class
